\name{bayespref-package}
\alias{bayespref-package}
\alias{bayespref}
\docType{package}
\title{
Hierarchical Bayesian analysis of ecological count data
}
\description{
This program implements a hierarchical Bayesian analysis of count data, such as preference experiments. It provides population-level and individual-level preference parameter estimates obtained via MCMC. It also allows for model comparison using Deviance Information Criterion.
}
\details{
\tabular{ll}{
Package: \tab bayespref\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-05-16\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

For information regarding functions implemented in \code{bayespref}, please see the appropriate function help pages.

}
\author{

James A. Fordyce \email{jfordyce@utk.edu} and Zachariah Gompert
\email{zgompert@uwyo.edu}

}

\keyword{ package }